Refactored parsing of string format into utility class.
